feat: Add SuggestionBuilder type and move SuggestInfo to index.ts
