+++ src/Suggestor/Suggestor.ts
@@ -4,23 +4,9 @@ import { DateParser } from '../Query/DateParser';
 import { doAutocomplete } from '../DateAbbreviations';
 import { Recurrence } from '../Recurrence';
 import { TaskRegularExpressions } from '../Task';
+import type { SuggestInfo } from '.';
 const { symbols } = TASK_FORMATS.tasksPluginEmoji.taskSerializer;
 const datePrefixCharacters = `${symbols.startDateSymbol}${symbols.scheduledDateSymbol}${symbols.dueDateSymbol}`;
-
-/*
- * A suggestion presented to the user and some metadata about it.
- */
-export type SuggestInfo = {
-    suggestionType?: 'match' | 'default' | 'empty';
-    // What to display to the user
-    displayText: string;
-    // What to append to the note
-    appendText: string;
-    // At what index in the line to do the insertion (if not specified, the cursor location is used)
-    insertAt?: number;
-    // How many characters to skip from the original line (e.g. if replacing existing text)
-    insertSkip?: number;
-};
 
 /*
  * Return a list of suggestions, either generic or more fine-grained to the words at the cursor.
